requirements.txt
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.35.0
tqdm>=4.65.0
datasets>=2.14.0
pyarrow>=12.0.0
torch>=2.0.0
